fix for the Oracle specific sql in find_object_id(): in Oracle, conditions testing for ROWNUM values greater than a positive integer are always false
"For example, this query returns no rows:
SELECT * FROM emp  WHERE ROWNUM > 1;
The first row fetched is assigned a ROWNUM of 1 and makes the condition false. The second row to be fetched is now the first row and is also assigned a ROWNUM of 1 and makes the condition false. All rows subsequently fail to satisfy the condition, so no rows are returned."